Update logic to increase file descriptor limit for a process to first check if new limit is greater than current limit before changing it when launching an environment for Android World.
This is important because the current limit does not strictly increase the limit but changes it to a fixed number (32768) in all cases.  When the current limit of file descriptors is above this, the call to increase the limit then actually decreases the number of file descriptors that can be created.

PiperOrigin-RevId: 784595147