Update implementation of function for pulling files from devices to create a new temporary directory for saving the file every time the function was called.
This ensures the function is thread safe. The old implementation would create a single temporary directory at the module level.  When running Android World from multiple threads, this could result in multiple instances trying to save files with the same name in the same directory, resulting in corrupt data.  The new implementation avoids this problem.

PiperOrigin-RevId: 836225493